[not verified] Update paid plan panel text based on plan availability
