+++ projects/plugins/jetpack/extensions/blocks/subscriptions/paid-plan.js
@@ -1,16 +1,32 @@
 import { PanelBody, PanelRow } from '@wordpress/components';
+import { compose } from '@wordpress/compose';
+import { withSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 import GetAddPaidPlanButton from './utils';
 
-export function PaidPlanPanel() {
+function PaidPlanPanel( { hasNewsletterPlans } ) {
 	const title = __( 'Paid Newsletter', 'jetpack' );
-	const text = __( 'Set up a paid plan for readers to access your content.', 'jetpack' );
+	const text = hasNewsletterPlans
+		? __( 'Manage paid plan for readers to access your content.', 'jetpack' )
+		: __( 'Set up paid plan for readers to access your content.', 'jetpack' );
+
 	return (
 		<PanelBody title={ title } initialOpen={ true }>
 			<PanelRow>{ text }</PanelRow>
 			<PanelRow>
-				<GetAddPaidPlanButton />
+				<GetAddPaidPlanButton hasNewsletterPlans={ hasNewsletterPlans } />
 			</PanelRow>
 		</PanelBody>
 	);
 }
+
+export default compose( [
+	withSelect( select => {
+		const newsletterPlans = select( 'jetpack/membership-products' )
+			?.getProducts()
+			?.filter( product => product.subscribe_as_site_subscriber );
+		return {
+			hasNewsletterPlans: newsletterPlans?.length !== 0,
+		};
+	} ),
+] )( PaidPlanPanel );
